+++ src/request-matcher/matchers/body.ts
@@ -1,9 +1,10 @@
 /**
  * Body matcher.
  */
-import isMatch from 'lodash/isMatch';
+import isMatchWith from 'lodash/isMatchWith';
 import { MockRequestSchema } from '../../protocol';
 import { MatchingContext } from '../context';
+import { isValueMatcher, matchValue } from '../value-matcher';
 
 export class BodyMatcher {
   constructor(private schema: MockRequestSchema) {}
@@ -54,7 +55,9 @@ export class BodyMatcher {
       return false;
     }
 
-    const result = isMatch(actualBodyParsed, expectedBody);
+    const result = isMatchWith(actualBodyParsed, expectedBody, (objValue, srcValue) =>
+      isValueMatcher(srcValue) ? matchValue(srcValue, objValue) : undefined,
+    );
     ctx.logger?.log(`body`, this.expectedBodyStr, trimLongString(actualBody));
 
     return result;