+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 > This project follows the [Keep a Changelog](https://keepachangelog.com) format.
 
 ## [Unreleased]
+* add `$contains` and `$regex` value matchers for request url, query, headers and body.
 
 ## [0.4.0] - 2026-06-02
 * refactor and export response builder.